@@ -1,68 +1,68 @@
 {
     "name": "open-capture-for-invoices",
-    "version": "2.0.5",
+    "version": "2.1.0",
     "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build-watch": "ng build --watch",
-        "build:prod": "NODE_ENV=production ng build --prod",
-        "reload-packages": "rm -Rf node_modules && rm package-lock.json && npm cache clean --force && npm i --force",
-        "extract-translations": "ngx-translate-extract --input ./ --output ./src/assets/i18n/frontend/*.json --clean --sort --format namespaced-json --marker _"
+        "build-watch": "ng build --watch --configuration development",
+        "build:prod": "NODE_ENV=production ng build --configuration production",
+        "reload-packages": "rm -Rf node_modules && rm -f package-lock.json && npm cache clean --force && npm i --force",
+        "extract-translations": "ngx-translate-extract --input ./src/frontend/ --output ./src/assets/i18n/frontend/*.json --clean --sort --format namespaced-json --marker _",
+        "lint": "ng lint"
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "~11.2.14",
-        "@angular/cdk": "^11.2.13",
-        "@angular/common": "~11.2.14",
-        "@angular/compiler": "~11.2.14",
-        "@angular/core": "~11.2.14",
-        "@angular/forms": "~11.2.14",
-        "@angular/material": "^11.2.13",
-        "@angular/material-moment-adapter": "^11.2.13",
-        "@angular/platform-browser": "~11.2.14",
-        "@angular/platform-browser-dynamic": "~11.2.14",
-        "@angular/router": "~11.2.14",
+        "@angular/animations": "~13.2.4",
+        "@angular/cdk": "^13.2.4",
+        "@angular/common": "~13.2.4",
+        "@angular/compiler": "~13.2.4",
+        "@angular/core": "~13.2.4",
+        "@angular/forms": "~13.2.4",
+        "@angular/material": "^13.2.4",
+        "@angular/material-moment-adapter": "^13.2.4",
+        "@angular/platform-browser": "~13.2.4",
+        "@angular/platform-browser-dynamic": "~13.2.4",
+        "@angular/router": "~13.2.4",
         "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
-        "@fortawesome/fontawesome-free": "^5.15.4",
-        "@ngx-translate/core": "^13.0.0",
-        "@ngx-translate/http-loader": "^6.0.0",
-        "@schematics/angular": "^11.2.15",
-        "@swimlane/ngx-charts": "17.0.0",
-        "@syncfusion/ej2": "^19.2.60",
-        "@types/pdfjs-dist": "2.1.7",
+        "@fortawesome/fontawesome-free": "^6.0.0",
+        "@ngx-translate/core": "^14.0.0",
+        "@ngx-translate/http-loader": "^7.0.0",
+        "@schematics/angular": "^13.2.5",
+        "@swimlane/ngx-charts": "^20.1.0",
+        "@syncfusion/ej2": "^19.4.53",
         "jquery": "^3.6.0",
         "moment": "^2.29.1",
-        "ng-click-outside": "^8.0.0",
-        "ng-sortgrid": "^5.2.1",
+        "ng-click-outside2": "^10.0.4",
+        "ng-sortgrid": "^5.3.0",
         "ng2-datepicker": "^12.0.0",
-        "ng2-pdfjs-viewer": "^6.0.2",
-        "ngx-file-drag-drop": "^4.0.4",
-        "ngx-mat-select-search": "^3.3.3",
-        "pdfjs-dist": "^2.9.359",
-        "postcss": "^8.4.5",
+        "ngx-file-drag-drop": "^5.1.3",
+        "ngx-mat-select-search": "^4.0.2",
         "remove-accents": "^0.4.2",
-        "rxjs": "~6.6.7",
-        "tailwindcss": "^2.2.19"
+        "rxjs": "~7.5.4"
     },
     "devDependencies": {
-        "@angular-builders/custom-webpack": "^11.1.1",
-        "@angular-devkit/build-angular": "^0.1102.14",
-        "@angular/cli": "~11.2.15",
-        "@angular/compiler-cli": "~11.2.14",
-        "@biesbjerg/ngx-translate-extract": "^7.0.4",
-        "autoprefixer": "^10.4.0",
-        "jasmine-core": "^3.10.1",
-        "jasmine-spec-reporter": "^7.0.0",
-        "karma": "~6.3.8",
-        "karma-chrome-launcher": "~3.1.0",
-        "karma-coverage": "~2.0.3",
-        "karma-jasmine": "~4.0.1",
-        "karma-jasmine-html-reporter": "^1.7.0",
-        "ngx-build-plus": "^11.0.0",
-        "protractor": "~7.0.0",
-        "ts-node": "^9.1.1",
-        "tslint": "~6.1.3",
-        "typescript": "~4.1.6",
-        "zone.js": "^0.11.4"
+        "@angular-devkit/build-angular": "^13.2.5",
+        "@angular-devkit/build-webpack": "^0.1302.5",
+        "@angular-eslint/builder": "13.1.0",
+        "@angular-eslint/eslint-plugin": "13.1.0",
+        "@angular-eslint/eslint-plugin-template": "13.1.0",
+        "@angular-eslint/schematics": "13.1.0",
+        "@angular-eslint/template-parser": "13.1.0",
+        "@angular/cli": "~13.2.5",
+        "@angular/compiler-cli": "~13.2.4",
+        "@bartholomej/ngx-translate-extract": "^8.0.2",
+        "@types/d3-scale": "^4.0.2",
+        "@types/d3-selection": "^3.0.2",
+        "@types/node": "^17.0.21",
+        "@typescript-eslint/eslint-plugin": "5.12.1",
+        "@typescript-eslint/parser": "5.12.1",
+        "autoprefixer": "^10.4.2",
+        "eslint": "^8.10.0",
+        "eslint-plugin-import": "latest",
+        "eslint-plugin-jsdoc": "latest",
+        "eslint-plugin-prefer-arrow": "latest",
+        "postcss": "^8.4.7",
+        "tailwindcss": "^3.0.23",
+        "ts-node": "^10.3.1",
+        "typescript": "~4.5.5",
+        "zone.js": "~0.11.4"
     }
 }
