@@ -1,34 +1,34 @@
 /** This file is part of Open-Capture for Invoices.
 
-Open-Capture for Invoices is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
+ Open-Capture for Invoices is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
 
-Open-Capture is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
+ Open-Capture is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU General Public License for more details.
 
-You should have received a copy of the GNU General Public License
-along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
+ You should have received a copy of the GNU General Public License
+ along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
 
-@dev : Nathan Cheval <nathan.cheval@outlook.fr>
-@dev : Oussama BRICH <oussama.brich@edissyum.com> */
+ @dev : Nathan Cheval <nathan.cheval@outlook.fr>
+ @dev : Oussama BRICH <oussama.brich@edissyum.com> */
 
 import { BrowserModule, Title } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { AppMaterialModule } from './app-material.module';
 import { ServicesModule } from '../services/services.module';
-import { MAT_FORM_FIELD_DEFAULT_OPTIONS } from "@angular/material/form-field";
-import { DragDropModule } from "@angular/cdk/drag-drop";
+import { MAT_FORM_FIELD_DEFAULT_OPTIONS } from '@angular/material/form-field';
+import { DragDropModule } from '@angular/cdk/drag-drop';
 
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { TranslateLoader, TranslateModule, TranslateService } from "@ngx-translate/core";
-import { HttpClient } from "@angular/common/http";
-import { HttpClientModule } from "@angular/common/http";
-import { TranslateHttpLoader } from "@ngx-translate/http-loader";
-import { FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { TranslateLoader, TranslateModule, TranslateService } from '@ngx-translate/core';
+import { HttpClient } from '@angular/common/http';
+import { HttpClientModule } from '@angular/common/http';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { AppRoutingModule } from './app-routing.module';
 import { SettingsRoutingModule } from './settings/settings-routing.module';
@@ -37,20 +37,18 @@ import { VerifierViewerComponent } from './verifier/viewer/verifier-viewer.compo
 import { VerifierListComponent } from './verifier/list/verifier-list.component';
 import { SplitterViewerComponent } from './splitter/viewer/splitter-viewer.component';
 import { SplitterListComponent } from './splitter/list/splitter-list.component';
-import { DocumentTypeComponent } from './settings/splitter/document-type/document-type.component';
 
 import { FilterPipe } from '../services/pipes/filter.pipe';
-import { NgxFileDragDropModule } from "ngx-file-drag-drop";
-import { ClickOutsideModule } from 'ng-click-outside';
-import { PdfJsViewerModule } from 'ng2-pdfjs-viewer';
-import { NgxMatSelectSearchModule } from "ngx-mat-select-search";
+import { NgxFileDragDropModule } from 'ngx-file-drag-drop';
+import { NgClickOutsideModule } from 'ng-click-outside2';
+import { NgxMatSelectSearchModule } from 'ngx-mat-select-search';
 import { NgsgModule } from 'ng-sortgrid';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
 
 import { DocumentTypeFactoryComponent } from './splitter/document-type-factory/document-type-factory.component';
-import { DocumentTreeComponent } from './splitter/document-tree/document-tree.component';
+import { DocumentTypeComponent } from './splitter/document-type/document-type.component';
 import { LoginComponent } from './login/login.component';
-import { LogoutComponent } from "./logout/logout.component";
+import { LogoutComponent } from './logout/logout.component';
 import { HomeComponent } from './home/home.component';
 import { MenuComponent } from './menu/menu.component';
 import { NotFoundComponent } from './not-found/not-found.component';
@@ -58,9 +56,9 @@ import { UserProfileComponent } from './profile/profile.component';
 import { UploadComponent } from './upload/upload.component';
 import { UsersListComponent } from './settings/general/users/list/users-list.component';
 import { SettingsComponent } from './settings/settings.component';
-import { MatPaginatorIntl } from "@angular/material/paginator";
-import { CustomMatPaginatorIntl } from "./custom-mat-paginator";
-import { LastUrlService } from "../services/last-url.service";
+import { MatPaginatorIntl } from '@angular/material/paginator';
+import { CustomMatPaginatorIntl } from './custom-mat-paginator';
+import { LastUrlService } from '../services/last-url.service';
 import { AboutUsComponent } from './settings/general/about-us/about-us.component';
 import { CreateUserComponent } from './settings/general/users/create/create-user.component';
 import { UpdateUserComponent } from './settings/general/users/update/update-user.component';
@@ -94,9 +92,13 @@ import { SplitterInputListComponent } from './settings/splitter/inputs/list/inpu
 import { SplitterUpdateOutputComponent } from './settings/splitter/output/update/update-output.component';
 import { SplitterCreateOutputComponent } from './settings/splitter/output/create/create-output.component';
 import { SplitterListOutputComponent } from './settings/splitter/output/list/list-output.component';
-import { SplitterFormBuilderComponent } from "./settings/splitter/form/builder/form-builder.component";
-import { SplitterFormListComponent } from "./settings/splitter/form/list/form-list.component";
+import { SplitterFormBuilderComponent } from './settings/splitter/form/builder/form-builder.component';
+import { SplitterFormListComponent } from './settings/splitter/form/list/form-list.component';
 import { StatisticsComponent } from './statistics/statistics.component';
+import { CreateDocTypeComponent } from './settings/splitter/doc-types/create/create-doc-type.component';
+import { ListDocTypeComponent } from './settings/splitter/doc-types/list/list-doc-type.component';
+import { CreateFolderDocTypeComponent } from './settings/splitter/doc-types/create-folder/create-folder-doc-type.component';
+import { ConfigurationsComponent } from './settings/general/configurations/configurations.component';
 
 export function createTranslateLoader(http: HttpClient) {
     return new TranslateHttpLoader(http, 'assets/i18n/frontend/', '.json');
@@ -111,7 +113,7 @@ export function createTranslateLoader(http: HttpClient) {
         SplitterListComponent,
         SplitterViewerComponent,
         DocumentTypeFactoryComponent,
-        DocumentTreeComponent,
+        DocumentTypeComponent,
         LoginComponent,
         LogoutComponent,
         HomeComponent,
@@ -129,7 +131,6 @@ export function createTranslateLoader(http: HttpClient) {
         CreateRoleComponent,
         LoaderComponent,
         CustomFieldsComponent,
-        DocumentTypeComponent,
         SeparatorComponent,
         FormBuilderComponent,
         FormListComponent,
@@ -158,7 +159,11 @@ export function createTranslateLoader(http: HttpClient) {
         SplitterCreateOutputComponent,
         SplitterListOutputComponent,
         HistoryComponent,
-        StatisticsComponent
+        StatisticsComponent,
+        CreateDocTypeComponent,
+        ListDocTypeComponent,
+        CreateFolderDocTypeComponent,
+        ConfigurationsComponent
     ],
     imports: [
         BrowserModule,
@@ -181,10 +186,9 @@ export function createTranslateLoader(http: HttpClient) {
         ReactiveFormsModule,
         FormsModule,
         NgxFileDragDropModule,
-        ClickOutsideModule,
+        NgClickOutsideModule,
         DragDropModule,
         NgxMatSelectSearchModule,
-        PdfJsViewerModule,
     ],
     providers: [
         Title,
