@@ -21,7 +21,7 @@ from flask_cors import CORS
 from flask_babel import Babel
 from flask import request, session, Flask
 from src.backend.import_rest import auth, locale, config, user, splitter, verifier, roles, privileges, custom_fields, \
-    forms, status, accounts, outputs, maarch, inputs, positions_masks, history
+    forms, status, accounts, outputs, maarch, inputs, positions_masks, history, doctypes
 
 
 app = Flask(__name__, instance_relative_config=True)
@@ -58,6 +58,7 @@ app.register_blueprint(verifier.bp)
 app.register_blueprint(privileges.bp)
 app.register_blueprint(custom_fields.bp)
 app.register_blueprint(positions_masks.bp)
+app.register_blueprint(doctypes.bp)
 
 
 @babel.localeselector
