@@ -15,11 +15,11 @@
 
 # @dev : Nathan Cheval <nathan.cheval@edissyum.com>
 
-import git
 import subprocess
 import configparser
 from flask_babel import gettext
 from flask import current_app, Blueprint
+from src.backend.import_models import config
 from src.backend.main import create_classes_from_current_config
 
 bp = Blueprint('dashboard', __name__)
@@ -46,10 +46,20 @@ def change_locale_in_config(lang):
         return {'errors': gettext("CHANGE_LOCALE_ERROR"), 'message': str(e)}, 500
 
 
-def get_current_git_version(cfg):
-    repo = git.Repo(cfg['GLOBAL']['projectpath'])
-    current_tag = next((tag for tag in repo.tags if tag.commit == repo.head.commit), None)
-    return str(current_tag)
+def retrieve_configurations(args):
+    custom_fields_res, error = config.retrieve_configurations(args)
+
+    if error is None:
+        response = {
+            "configurations": custom_fields_res
+        }
+        return response, 200
+
+    response = {
+        "errors": "RETRIEVE_CONFIGURATIONS_ERRORS",
+        "message": error
+    }
+    return response, 401
 
 
 def get_last_git_version():
