@@ -17,6 +17,7 @@
 # @dev : Oussama Brich <oussama.brich@edissyum.com>
 
 from src.backend.import_models import custom_fields
+from src.backend.import_controllers import forms
 
 
 def add_custom_field(args):
@@ -53,13 +54,24 @@ def retrieve_custom_fields(args):
 def update(args):
     res, error = custom_fields.update(args)
     if res:
-        response = {
-            "res": res
-        }
-        return response, 200
+        forms.update_custom_field_from_forms(args)
+        return '', 200
     else:
         response = {
             "errors": "CUSTOM_FIELDS_ERROR",
+            "message": error
+        }
+        return response, 401
+
+
+def delete(args):
+    res, error = custom_fields.delete(args)
+    if not error:
+        forms.delete_custom_field_from_forms(args)
+        return '', 200
+    else:
+        response = {
+            "errors": "CUSTOM_FIELDS_DELETE_ERROR",
             "message": error
         }
         return response, 401
