@@ -14,20 +14,21 @@
 # along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
 
 # @dev : Nathan Cheval <nathan.cheval@outlook.fr>
-
+import json
 import os
 import sys
 import time
 import tempfile
+from symbol import import_from
+
 from kuyruk import Kuyruk
-from src.backend.import_process import OCForInvoices_splitter
 from src.backend.main import timer, check_file, create_classes
-from src.backend.import_classes import _Files, _Config, _Splitter, _SeparatorQR, _Log, _Mail
+from src.backend.import_classes import _Files, _Config, _Splitter, _SeparatorQR, _Log
 
 OCforInvoices = Kuyruk()
 
 
-# @OCforInvoices.task(queue='splitter')
+@OCforInvoices.task(queue='splitter')
 def launch(args):
     start = time.time()
 
@@ -42,7 +43,19 @@ def launch(args):
     tmp_folder = tempfile.mkdtemp(dir=config.cfg['SPLITTER']['batchpath']) + '/'
     filename = tempfile.NamedTemporaryFile(dir=tmp_folder).name
     files = _Files(filename, log, locale, config)
-    separator_qr = _SeparatorQR(log, config, tmp_folder, 'splitter', files)
+
+    remove_blank_pages = False
+    if 'input_id' in args:
+        input_settings = database.select({
+            'select': ['*'],
+            'table': ['inputs'],
+            'where': ['input_id = %s', 'module = %s'],
+            'data': [args['input_id'], 'verifier'],
+        })
+        if input_settings:
+            remove_blank_pages = input_settings[0]['remove_blank_pages']
+
+    separator_qr = _SeparatorQR(log, config, tmp_folder, 'splitter', files, remove_blank_pages)
     splitter = _Splitter(config, database, locale, separator_qr, log)
 
     if args.get('isMail') is not None and args['isMail'] is True:
@@ -53,7 +66,35 @@ def launch(args):
     if args['file'] is not None:
         path = args['file']
         if check_file(files, path, config, log) is not False:
-            OCForInvoices_splitter.process(args, path, log, splitter, files, tmp_folder, config)
+            splitter_method = database.select({
+                'select': ['splitter_method_id'],
+                'table': ['inputs'],
+                'where': ['status <> %s', 'input_id = %s', 'module = %s'],
+                'data': ['DEL', args['input_id'], 'splitter']
+            })[0]
+            available_split_methods_path = config.cfg['SPLITTER']['methodspath'] + "/splitter_methods.json"
+            if len(splitter_method) > 0 and os.path.isfile(available_split_methods_path):
+                with open(available_split_methods_path, encoding='utf-8') as json_file:
+                    available_split_methods = json.load(json_file)
+                    for available_split_method in available_split_methods['methods']:
+                        if available_split_method['id'] == splitter_method['splitter_method_id']:
+                            split_method = import_from(config, available_split_method['script'], available_split_method['method'])
+                            log.info('Split using method : {}'.format(available_split_method['id']))
+                            split_method(args, path, log, splitter, files, tmp_folder, config)
     database.conn.close()
     end = time.time()
     log.info('Process end after ' + timer(start, end) + '')
+
+
+def import_from(config, script, method):
+    """
+    Import an attribute, function or class from a module.
+    :param method: Method to call
+    :param path: A path descriptor in the form of 'pkg.module.submodule:attribute'
+    :type path: str
+    """
+    import sys
+    sys.path.append(config.cfg['SPLITTER']['methodspath'])
+    script = script.replace('.py', '')
+    module = __import__(script, fromlist=method)
+    return getattr(module, method)
