@@ -44,8 +44,7 @@ def suppliers_list():
             "LOWER(siren) LIKE '%%" + request.args['search'].lower() + "%%' OR "
             "LOWER(vat_number) LIKE '%%" + request.args['search'].lower() + "%%'"
         )
-        args['offset'] = ''
-        args['limit'] = ''
+
     res = accounts.retrieve_suppliers(args)
     return make_response(res[0], res[1])
 
@@ -177,8 +176,6 @@ def customers_list():
             "LOWER(siren) LIKE '%%" + request.args['search'].lower() + "%%' OR "
             "LOWER(vat_number) LIKE '%%" + request.args['search'].lower() + "%%'"
         )
-        args['offset'] = ''
-        args['limit'] = ''
 
     res = accounts.retrieve_customers(args)
     return make_response(res[0], res[1])
