@@ -15,7 +15,7 @@
 
 # @dev : Nathan Cheval <nathan.cheval@edissyum.com>
 
-from flask import Blueprint, request, jsonify, make_response
+from flask import Blueprint, jsonify, make_response, request
 from src.backend.import_controllers import auth, config
 from src.backend.main import create_classes_from_current_config
 
@@ -25,16 +25,32 @@ bp = Blueprint('config', __name__,  url_prefix='/ws/')
 @bp.route('config/readConfig', methods=['GET'])
 @auth.token_required
 def read_config():
-    if request.method == 'GET':
-        _vars = create_classes_from_current_config()
-        return make_response(jsonify({'config': _vars[1].cfg})), 200
+    _vars = create_classes_from_current_config()
+    return make_response(jsonify({'config': _vars[1].cfg})), 200
+
+
+@bp.route('config/getConfigurations', methods=['GET'])
+@auth.token_required
+def get_configurations():
+    args = {
+        'select': ['*', 'count(*) OVER() as total'],
+        'where': [],
+        'offset': request.args['offset'] if 'offset' in request.args else '',
+        'limit': request.args['limit'] if 'limit' in request.args else ''
+    }
+
+    if 'search' in request.args and request.args['search']:
+        args['where'].append(
+            "LOWER(label) LIKE '%%" + request.args['search'].lower() + "%%' OR "
+                                                                       "LOWER(data ->> 'description') LIKE '%%" + request.args['search'].lower() + "%%'"
+        )
+    res = config.retrieve_configurations(args)
+    return make_response(jsonify(res[0])), res[1]
 
 
 @bp.route('config/gitInfo', methods=['GET'])
 @auth.token_required
 def get_git_info():
-    _vars = create_classes_from_current_config()
     return make_response({
-        'git_current': config.get_current_git_version(_vars[1].cfg),
         'git_latest': config.get_last_git_version()
     }), 200
