@@ -1,19 +1,19 @@
 /** This file is part of Open-Capture for Invoices.
 
-Open-Capture for Invoices is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
+ Open-Capture for Invoices is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
 
-Open-Capture is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
+ Open-Capture is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU General Public License for more details.
 
-You should have received a copy of the GNU General Public License
-along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
+ You should have received a copy of the GNU General Public License
+ along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
 
-@dev : Nathan Cheval <nathan.cheval@outlook.fr> */
+ @dev : Nathan Cheval <nathan.cheval@outlook.fr> */
 
 import {MatPaginatorIntl} from "@angular/material/paginator";
 import {Injectable} from '@angular/core';
@@ -85,7 +85,7 @@ export class CustomMatPaginatorIntl extends MatPaginatorIntl {
         }
     }
 
-    getRangeLabel = (page: number, pageSize: number, length: number) =>  {
+    override getRangeLabel = (page: number, pageSize: number, length: number) =>  {
         if (length === 0 || pageSize === 0) { return '0 ' + this.translate.instant('PAGINATOR.of') + ` ${length}`; }
 
         length = Math.max(length, 0);
@@ -95,10 +95,10 @@ export class CustomMatPaginatorIntl extends MatPaginatorIntl {
         // If the start index exceeds the list length, do not try and fix the end index to the end.
         const endIndex = startIndex < length ?
             Math.min(startIndex + pageSize, length):
-                startIndex + pageSize;
+            startIndex + pageSize;
 
         const nbPage = Math.ceil(length / pageSize);
         return ` ${startIndex + 1} - ${endIndex} ` + this.translate.instant('PAGINATOR.on') + ` ${length} ` + ' | ' +
             this.translate.instant('PAGINATOR.page') + ` ${page + 1} / ${nbPage}`;
-    }
+    };
 }
