@@ -1,25 +1,25 @@
 /** This file is part of Open-Capture for Invoices.
 
-Open-Capture for Invoices is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
+ Open-Capture for Invoices is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
 
-Open-Capture is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
+ Open-Capture is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU General Public License for more details.
 
-You should have received a copy of the GNU General Public License
-along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
+ You should have received a copy of the GNU General Public License
+ along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
 
-@dev : Nathan Cheval <nathan.cheval@outlook.fr> */
+ @dev : Nathan Cheval <nathan.cheval@outlook.fr> */
 
 import { Component, OnInit } from '@angular/core';
 import { SettingsService } from "../../../../services/settings.service";
 import { Router } from "@angular/router";
 import { PrivilegesService } from "../../../../services/privileges.service";
-import {API_URL} from "../../../env";
+import {API_URL, environment} from "../../../env";
 import {catchError, finalize, tap} from "rxjs/operators";
 import {of} from "rxjs";
 import {HttpClient} from "@angular/common/http";
@@ -36,6 +36,7 @@ export class AboutUsComponent implements OnInit {
     loading        : boolean   = true;
     gitVersion     : any       = 'dev';
     lastVersion    : any       = '';
+    isProd         : boolean   = false;
 
     constructor(
         public router: Router,
@@ -47,10 +48,11 @@ export class AboutUsComponent implements OnInit {
     ) {}
 
     ngOnInit(): void {
+        this.serviceSettings.init();
+        this.isProd = environment.production;
+        this.gitVersion = environment.VERSION;
         this.http.get(API_URL + '/ws/config/gitInfo', {headers: this.authService.headers}).pipe(
             tap((data: any) => {
-                if (data.git_current && data.git_current !== 'None')
-                    this.gitVersion = data.git_current;
                 if (data.git_latest)
                     this.lastVersion = data.git_latest;
             }),
