@@ -1,19 +1,19 @@
 /** This file is part of Open-Capture for Invoices.
 
-Open-Capture for Invoices is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
+ Open-Capture for Invoices is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
 
-Open-Capture is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
+ Open-Capture is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU General Public License for more details.
 
-You should have received a copy of the GNU General Public License
-along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
+ You should have received a copy of the GNU General Public License
+ along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
 
-@dev : Nathan Cheval <nathan.cheval@outlook.fr> */
+ @dev : Nathan Cheval <nathan.cheval@outlook.fr> */
 
 import {Component, OnInit} from '@angular/core';
 import {Router} from "@angular/router";
@@ -30,7 +30,6 @@ export class SettingsComponent implements OnInit {
     isMenuOpen              : boolean   = this.serviceSettings.getIsMenuOpen();
     selectedSetting         : string    = this.serviceSettings.getSelectedSetting();
     selectedParentSetting   : string    = this.serviceSettings.getSelectedParentSetting();
-    settingListOpenState    : boolean   = this.serviceSettings.getSettingListOpenState();
     settings                : any       = this.serviceSettings.getSettings();
 
     constructor(
@@ -38,7 +37,7 @@ export class SettingsComponent implements OnInit {
         public userService: UserService,
         public serviceSettings: SettingsService,
         public privilegesService: PrivilegesService
-    ) { }
+    ) {}
 
     ngOnInit(): void {
         this.serviceSettings.init();
@@ -55,9 +54,8 @@ export class SettingsComponent implements OnInit {
             }
         });
 
-        if (!settingsFound) {
-            this.router.navigateByUrl('/settings/general/users').then();
-        }
+        if (!settingsFound)
+            this.router.navigateByUrl('/settings/general/configurations').then();
     }
 
     toggleMenu() {
