@@ -1,20 +1,20 @@
 /** This file is part of Open-Capture for Invoices.
 
-Open-Capture for Invoices is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
+ Open-Capture for Invoices is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
 
-Open-Capture is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
+ Open-Capture is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
 
-You should have received a copy of the GNU General Public License
-along with Open-Capture for Invoices.  If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
+ You should have received a copy of the GNU General Public License
+ along with Open-Capture for Invoices.  If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
 
-@dev : Nathan Cheval <nathan.cheval@outlook.fr>
-@dev : Oussama Brich <oussama.brich@edissyum.com> */
+ @dev : Nathan Cheval <nathan.cheval@outlook.fr>
+ @dev : Oussama Brich <oussama.brich@edissyum.com> */
 
 import { Component, OnInit } from '@angular/core';
 import {HttpClient, HttpHeaders} from "@angular/common/http";
@@ -32,9 +32,9 @@ import {catchError, finalize, tap} from "rxjs/operators";
 import {of} from "rxjs";
 
 @Component({
-  selector: 'app-splitter-update-input',
-  templateUrl: './update-input.component.html',
-  styleUrls: ['./update-input.component.scss']
+    selector: 'app-splitter-update-input',
+    templateUrl: './update-input.component.html',
+    styleUrls: ['./update-input.component.scss']
 })
 export class SplitterUpdateInputComponent implements OnInit {
     headers         : HttpHeaders   = this.authService.headers;
@@ -71,6 +71,20 @@ export class SplitterUpdateInputComponent implements OnInit {
             type: 'select',
             control: new FormControl(),
             required: true,
+        },
+        {
+            id: 'splitter_method_id',
+            label: this.translate.instant('INPUT.splitter_method'),
+            type: 'select',
+            control: new FormControl(),
+            required: true,
+            values: [],
+        },
+        {
+            id: 'remove_blank_pages',
+            label: this.translate.instant('INPUT.remove_blank_pages'),
+            type: 'boolean',
+            control: new FormControl()
         },
     ];
 
@@ -123,6 +137,26 @@ export class SplitterUpdateInputComponent implements OnInit {
                 return of(false);
             })
         ).subscribe();
+        this.inputForm.forEach(element => {
+            if (element.id === 'splitter_method_id') {
+                this.http.get(API_URL + '/ws/splitter/methods', {headers: this.authService.headers}).pipe(
+                    tap((data: any) => {
+                        data.split_methods.forEach((option: any) => {
+                            element.values.push({
+                                id      : option.id,
+                                label   : option.label,
+                            });
+                        });
+                    }),
+                    finalize(() => this.loading = false),
+                    catchError((err: any) => {
+                        console.debug(err);
+                        this.notify.handleErrors(err);
+                        return of(false);
+                    })
+                ).subscribe();
+            }
+        });
     }
 
     isValidForm() {
