@@ -1,20 +1,20 @@
 /** This file is part of Open-Capture for Invoices.
 
-Open-Capture for Invoices is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
+ Open-Capture for Invoices is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
 
-Open-Capture is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
+ Open-Capture is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
 
-You should have received a copy of the GNU General Public License
-along with Open-Capture for Invoices.  If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
+ You should have received a copy of the GNU General Public License
+ along with Open-Capture for Invoices.  If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
 
-@dev : Nathan Cheval <nathan.cheval@outlook.fr>
-@dev : Oussama Brich <oussama.brich@edissyum.com> */
+ @dev : Nathan Cheval <nathan.cheval@outlook.fr>
+ @dev : Oussama Brich <oussama.brich@edissyum.com> */
 
 import { Component, OnInit } from '@angular/core';
 import {FormControl} from "@angular/forms";
@@ -29,15 +29,17 @@ import {PrivilegesService} from "../../../../../services/privileges.service";
 import {API_URL} from "../../../../env";
 import {catchError, finalize, tap} from "rxjs/operators";
 import {of} from "rxjs";
+import {marker} from "@biesbjerg/ngx-translate-extract-marker";
 
 @Component({
-  selector: 'app-splitter-create-output',
-  templateUrl: './create-output.component.html',
-  styleUrls: ['./create-output.component.scss']
+    selector: 'app-splitter-create-output',
+    templateUrl: './create-output.component.html',
+    styleUrls: ['./create-output.component.scss']
 })
 export class SplitterCreateOutputComponent implements OnInit {
     loading             : boolean       = true;
     outputsTypes        : any[]         = [];
+    selectedOutputType  : any;
     outputForm          : any[]         = [
         {
             id: 'output_type_id',
@@ -53,7 +55,40 @@ export class SplitterCreateOutputComponent implements OnInit {
             type: 'text',
             control: new FormControl(),
             required: true,
-        }
+        },
+        {
+            id: 'compress_type',
+            label: this.translate.instant('OUTPUT.compress_type'),
+            type: 'select',
+            control: new FormControl(),
+            values: [
+                {
+                    'id': '',
+                    'label': marker('OUTPUT.no_compress')
+                },
+                {
+                    'id': 'screen',
+                    'label': marker('OUTPUT.compress_screen')
+                },
+                {
+                    'id': 'ebook',
+                    'label': marker('OUTPUT.compress_ebook')
+                },
+                {
+                    'id': 'prepress',
+                    'label': marker('OUTPUT.compress_prepress')
+                },
+                {
+                    'id': 'printer',
+                    'label': marker('OUTPUT.compress_printer')
+                },
+                {
+                    'id': 'default',
+                    'label': marker('OUTPUT.compress_default')
+                }
+            ],
+            required: false,
+        },
     ];
 
     constructor(
@@ -68,6 +103,14 @@ export class SplitterCreateOutputComponent implements OnInit {
     ) {}
 
     ngOnInit(): void {
+        this.serviceSettings.init();
+
+        this.outputForm.forEach((element: any) => {
+            if (element.id === 'compress_type') {
+                element.control.setValue('');
+            }
+        });
+
         this.http.get(API_URL + '/ws/outputs/getOutputsTypes?module=splitter', {headers: this.authService.headers}).pipe(
             tap((data: any) => {
                 this.outputsTypes = data.outputs_types;
@@ -97,13 +140,15 @@ export class SplitterCreateOutputComponent implements OnInit {
         if (this.isValidForm(this.outputForm)) {
             const outputTypeId = this.getValueFromForm(this.outputForm, 'output_type_id');
             const outputLabel = this.getValueFromForm(this.outputForm, 'output_label');
+            const compressType = this.getValueFromForm(this.outputForm, 'compress_type');
             this.http.post(API_URL + '/ws/outputs/create',
                 {
                     'args': {
-                    'output_type_id': outputTypeId,
-                    'output_label'  : outputLabel,
-                    'module'        : 'splitter',
-                }}, {headers: this.authService.headers},
+                        'output_type_id': outputTypeId,
+                        'output_label'  : outputLabel,
+                        'compress_type' : compressType,
+                        'module'        : 'splitter',
+                    }}, {headers: this.authService.headers},
             ).pipe(
                 tap((data: any) => {
                     this.notify.success(this.translate.instant('OUTPUT.created'));
@@ -140,4 +185,7 @@ export class SplitterCreateOutputComponent implements OnInit {
         return error;
     }
 
+    changeOutputType(event: any) {
+        this.selectedOutputType = event.value;
+    }
 }
