@@ -33,7 +33,7 @@ import {FileValidators} from "ngx-file-drag-drop";
 import {DomSanitizer} from "@angular/platform-browser";
 import {ConfigService} from "../../../../../services/config.service";
 import {HistoryService} from "../../../../../services/history.service";
-declare var $: any;
+declare const $: any;
 
 @Component({
     selector: 'update-positions-mask',
