@@ -1,3 +1,20 @@
+/** This file is part of Open-Capture for Invoices.
+
+ Open-Capture for Invoices is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ Open-Capture is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
+
+ @dev : Nathan CHEVAL <nathan.cheval@edissyum.com> */
+
 import { Component, OnInit } from '@angular/core';
 import {API_URL} from "../env";
 import {catchError, tap} from "rxjs/operators";
@@ -7,7 +24,7 @@ import {AuthService} from "../../services/auth.service";
 import {NotificationService} from "../../services/notifications/notifications.service";
 import {SettingsService} from "../../services/settings.service";
 import {TranslateService} from "@ngx-translate/core";
-import {MAT_FORM_FIELD_DEFAULT_OPTIONS} from "@angular/material/form-field";
+import {marker} from "@biesbjerg/ngx-translate-extract-marker";
 
 @Component({
     selector: 'app-statistics',
@@ -22,10 +39,30 @@ export class StatisticsComponent implements OnInit {
     options : any = [
         {
             'id': 'invoices_validated_per_user',
-            'label': 'STATISTICS.invoices_validated_per_user'
+            'label': this.translate.instant('STATISTICS.invoices_validated_per_user')
         }
     ];
+    diagramTypes : any = [
+        {
+            'id': 'vertical-bar',
+            'label': marker('STATISTICS.diagram_vertical_bar'),
+            'logo': 'fa-chart-column'
+        },
+        {
+            'id': 'pie-chart',
+            'label': marker('STATISTICS.diagram_pie_chart'),
+            'logo': 'fa-chart-pie'
+        },
+        {
+            'id': 'pie-grid',
+            'label': marker('STATISTICS.diagram_pie_grid'),
+            'logo': 'fa-grip'
+        },
+
+    ];
+
     selectedStatistic : any;
+    selectedDiagramType : string = 'vertical-bar';
 
     constructor(
         private http: HttpClient,
@@ -66,7 +103,6 @@ export class StatisticsComponent implements OnInit {
                                         'value': historyCpt
                                     });
                                 });
-                                console.log(this.data);
                                 this.loading = false;
                             }),
                             catchError((err: any) => {
@@ -96,6 +132,16 @@ export class StatisticsComponent implements OnInit {
             this.options.forEach((option: any) => {
                 if (option.id === event.value) {
                     this.selectedStatistic = option;
+                }
+            });
+        }
+    }
+
+    changeDiagramType(event: any) {
+        if (event.value) {
+            this.diagramTypes.forEach((option: any) => {
+                if (option.id === event.value) {
+                    this.selectedDiagramType = option.id;
                 }
             });
         }
