@@ -1,19 +1,19 @@
 /** This file is part of Open-Capture for Invoices.
 
-Open-Capture for Invoices is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
+ Open-Capture for Invoices is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
 
-Open-Capture is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
+ Open-Capture is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU General Public License for more details.
 
-You should have received a copy of the GNU General Public License
-along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
+ You should have received a copy of the GNU General Public License
+ along with Open-Capture for Invoices. If not, see <https://www.gnu.org/licenses/gpl-3.0.html>.
 
-@dev : Nathan Cheval <nathan.cheval@outlook.fr> */
+ @dev : Nathan Cheval <nathan.cheval@outlook.fr> */
 
 import {Component, OnInit} from '@angular/core';
 import {DomSanitizer, SafeUrl} from '@angular/platform-browser';
@@ -30,12 +30,10 @@ import { FormControl} from "@angular/forms";
 import { DatePipe } from '@angular/common';
 import { LocalStorageService } from "../../../services/local-storage.service";
 import { ConfigService } from "../../../services/config.service";
-import 'moment/locale/en-gb';
-import 'moment/locale/fr';
 import * as moment from 'moment';
 import { UserService } from "../../../services/user.service";
 import { HistoryService } from "../../../services/history.service";
-declare var $: any;
+declare const $: any;
 
 
 @Component({
@@ -87,8 +85,8 @@ export class VerifierViewerComponent implements OnInit {
             'label': marker('FORMS.supplier')
         },
         {
-            'id': 'facturation',
-            'label': marker('FACTURATION.facturation')
+            id: 'facturation',
+            label: marker('FACTURATION.facturation')
         },
         {
             'id': 'other',
@@ -142,7 +140,7 @@ export class VerifierViewerComponent implements OnInit {
         this.currentFilename = this.invoice.full_jpg_filename;
         await this.getThumb(this.invoice.full_jpg_filename);
         if (this.invoice.form_id) {
-           await this.generateOutputs(this.invoice.form_id);
+            await this.generateOutputs(this.invoice.form_id);
         }
 
         this.formList = await this.getAllForm();
@@ -218,7 +216,7 @@ export class VerifierViewerComponent implements OnInit {
             }
         }
         if (this.formSettings.supplier_verif && !this.token) {
-            let token: any;
+            const token: any;
             token = await this.generateTokenInsee();
             if (token['token']) {
                 if (token['token'].includes('ERROR')) {
@@ -402,7 +400,8 @@ export class VerifierViewerComponent implements OnInit {
                 const _field = this.form[category][this.form[category].length - 1];
 
                 if (field.id === 'accounting_plan') {
-                    let array = await this.retrieveAccountingPlan();
+                    let array : any = {};
+                    array = await this.retrieveAccountingPlan();
                     this.accountingPlanEmpty = Object.keys(array).length === 0;
                     if (this.accountingPlanEmpty) {
                         array = await this.retrieveDefaultAccountingPlan();
@@ -678,14 +677,14 @@ export class VerifierViewerComponent implements OnInit {
         this.http.put(API_URL + '/ws/verifier/invoices/' + this.invoice.id + '/updatePosition',
             {'args': {[this.lastId]: position}},
             {headers: this.authService.headers}).pipe(
-                tap(() => {
-                    this.invoice.positions[this.lastId] = position;
-                }),
-                catchError((err: any) => {
-                    console.debug(err);
-                    this.notify.handleErrors(err);
-                    return of(false);
-                })
+            tap(() => {
+                this.invoice.positions[this.lastId] = position;
+            }),
+            catchError((err: any) => {
+                console.debug(err);
+                this.notify.handleErrors(err);
+                return of(false);
+            })
         ).subscribe();
     }
 
@@ -705,14 +704,14 @@ export class VerifierViewerComponent implements OnInit {
         this.http.put(API_URL + '/ws/verifier/invoices/' + this.invoice.id + '/updatePage',
             {'args': {[this.lastId]: page}},
             {headers: this.authService.headers}).pipe(
-                tap(() => {
-                    this.invoice.pages[this.lastId] = page;
-                }),
-                catchError((err: any) => {
-                    console.debug(err);
-                    this.notify.handleErrors(err);
-                    return of(false);
-                })
+            tap(() => {
+                this.invoice.pages[this.lastId] = page;
+            }),
+            catchError((err: any) => {
+                console.debug(err);
+                this.notify.handleErrors(err);
+                return of(false);
+            })
         ).subscribe();
     }
 
@@ -986,7 +985,9 @@ export class VerifierViewerComponent implements OnInit {
             return value.length === size;
         }
 
-        if (isNaN(value) || value.length !== size) return false;
+        if (isNaN(value) || value.length !== size) {
+            return false;
+        }
         let bal     = 0;
         let total   = 0;
         for (let i = size - 1; i >= 0; i--) {
@@ -1062,7 +1063,9 @@ export class VerifierViewerComponent implements OnInit {
                 }
             });
         }
-        if (!valid) return;
+        if (!valid) {
+            return;
+        }
         this.saveData(arrayData);
         /*
             Executer les actions paramétrées dans les réglages du formulaires
@@ -1177,7 +1180,9 @@ export class VerifierViewerComponent implements OnInit {
                         }
                         this.saveInfo = false;
                         if (field) {
-                            if (parseInt(String(page)) === this.currentPage) this.drawPositionByField(field, position);
+                            if (parseInt(String(page)) === this.currentPage) {
+                                this.drawPositionByField(field, position);
+                            }
                         }
                     }
                 }
@@ -1276,17 +1281,17 @@ export class VerifierViewerComponent implements OnInit {
                     ).subscribe();
                 } else {
                     this.form['supplier'].forEach((element: any) => {
-                    if (element.id === 'vat_number') {
-                        setTimeout(() => {
-                            if (!this.token) {
-                                element.control.setErrors({'vat_error': this.translate.instant('ERROR.ecu_api_not_up')});
-                            } else {
-                                element.control.setErrors({'vat_error': this.translate.instant('ERROR.wrong_vat_number_format')});
-                            }
-                            element.control.markAsTouched();
-                        }, 100);
-                    }
-                });
+                        if (element.id === 'vat_number') {
+                            setTimeout(() => {
+                                if (!this.token) {
+                                    element.control.setErrors({'vat_error': this.translate.instant('ERROR.ecu_api_not_up')});
+                                } else {
+                                    element.control.setErrors({'vat_error': this.translate.instant('ERROR.wrong_vat_number_format')});
+                                }
+                                element.control.markAsTouched();
+                            }, 100);
+                        }
+                    });
                 }
             }
         }
